uncomment cd4pe module
I want to include this module in the modulepath when this control-repo is deployed onto the master. Then the profile class will wrap around/instantiate the cd4pe class that is part of this module that I have included here. Puppetfile will have ALL component modules that you will use.